# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Blockyard is a platform for deploying and managing Shiny applications.",
        "title": "Blockyard API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/apps/{id}/access": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all ACL entries for an app. Requires owner or admin role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "access"
                ],
                "summary": "List access grants",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID (UUID) or name",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.accessGrantResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Grant a user viewer or collaborator access to an app. Requires owner or admin role.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "access"
                ],
                "summary": "Grant access",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID (UUID) or name",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Access grant",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.grantRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Access granted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            }
        },
        "/apps/{id}/access/{kind}/{principal}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a specific access grant from an app. Requires owner or admin role.",
                "tags": [
                    "access"
                ],
                "summary": "Revoke access",
                "parameters": [
                    {
                        "type": "string",
                        "description": "App ID (UUID) or name",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Grant kind (e.g. 'user')",
                        "name": "kind",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User sub",
                        "name": "principal",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Access revoked"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.accessGrantResponse": {
            "type": "object",
            "properties": {
                "granted_at": {
                    "type": "string"
                },
                "granted_by": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "principal": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_api.grantRequest": {
            "type": "object",
            "properties": {
                "kind": {
                    "description": "\"user\"",
                    "type": "string"
                },
                "principal": {
                    "type": "string"
                },
                "role": {
                    "description": "\"viewer\" | \"collaborator\"",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Personal access token (PAT). Format: \"Bearer by_...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}